Moved tooltip out of popup
