+++ packages/react-components/src/date-picker/tests/jest/single-date-picker.test.jsx
@@ -15,7 +15,7 @@ jest.mock("@react-components/date-picker/src/react-dates-wrapper.jsx", () => {
     };
 });
 
-jest.mock("@react-components/popup/src/popup/fade-in.jsx", () => {
+jest.mock("@react-components/popup/src/fade-in.jsx", () => {
     return {
         FadeIn: ({ active, children, className }) => {
             return (
