+++ packages/react-components/src/popup/src/index.js
@@ -1,2 +1,2 @@
 export * from "./popup";
-export * from "./tooltip";
+export * from "./positions";
