--- packages/react-components/src/popup/src/popup/index.js
@@ -1,2 +1,0 @@
-export * from "./popup";
-export * from "./positions";
