+++ packages/react-components/src/popup/stories/components/controlled-red-box.jsx
