+++ packages/react-components/src/popup/stories/components/red-box-popup.jsx
