+++ packages/react-components/src/popup/stories/popup.stories.mdx
@@ -9,7 +9,7 @@ import Imports from "./imports.mdx";
 <Meta
     title={createComponentSection("Popup")}
     parameters={paramsBuilder()
-        .docsLayout(<ReactComponentsLayout sourcePath="/packages/react-components/src/popup/src/popup" />)
+        .docsLayout(<ReactComponentsLayout sourcePath="/packages/react-components/src/popup/src" />)
         .build()}
 />
