+++ packages/react-components/src/tooltip/index.js
@@ -1,0 +1,1 @@
+export * from "./src";
