+++ packages/react-components/src/tooltip/stories/tooltip.stories.mdx
@@ -3,7 +3,7 @@ import { Button } from "@react-components/button";
 import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Label } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { PureTooltip, Tooltip } from "@react-components/popup";
+import { PureTooltip, Tooltip } from "@react-components/tooltip";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
